Разделение кода файла "main.js" на модули
